+++ pkg/controllers/proxy/graceful_termination_controller.go
@@ -1,4 +1,4 @@
-package graceful
+package proxy
 
 import (
 	"context"
